requirements.txt
albumentations==0.5.2
einops==0.3.2
gala==1.5
numpy==1.19.2
onnx==1.10.2
onnxruntime==1.10.0
opencv_python_headless==4.5.2.52
paddle==1.0.2
paddlepaddle_gpu==2.2.1
Pillow==9.0.1
scikit_image==0.18.1
skimage==0.0
torch==1.8.1
torchvision==0.9.1
tqdm==4.60.0
x2paddle==1.3.5
xlwt==1.3.0
